fix restore. table matcher regex, lzo file decompression.
